awesome : open science ajout numfocus
The mission of NumFOCUS is to promote open practices in research, data,
and scientific computing by serving as a fiscal sponsor for open source
projects and organizing community-driven educational programs.

Travis Oliphant (author of NumPy), Fernando Pérez (author of IPython),
Perry Greenfield (author of Numarray and Astropy), John Hunter (author
of Matplotlib), Jarrod Millman (release manager for SciPy), and Anthony
Scopatz (who came up with the name “NumFOCUS”) became the founding board
of NumFOCUS. Leah Silen was selected as the founding Executive Director.
In fall of 2012, NumFOCUS received 501(c)(3) public charity status as a
nonprofit in the United States.