awesome : ajout openstax
As an educational initiative, it's our mission to improve educational
access and learning for everyone. Through our partnerships with
philanthropic foundations and our alliance with other educational
resource companies, we're breaking down the most common barriers to
learning. Because we believe that everyone should and can have access
to knowledge.