awesome : ajout livre the sharing economy in Europe
This open access book considers the development of the sharing and
collaborative economy with a European focus, mapping across economic
sectors, and country-specific case studies. It looks at the roles the
sharing economy plays in sharing and redistribution of goods and
services across the population in order to maximise their functionality,
monetary exchange, and other aspects important to societies. It also
looks at the place of the sharing economy among various policies and how
the contexts of public policies, legislation, digital platforms, and
other infrastructure interrelate with the development and function of
the sharing economy.