awesome : ajout EU-FOSSA 2 - Free and Open Source Software Auditing
The EU-FOSSA project – short for Free and Open Source Software Auditing
– aims to increase the security and integrity of critical open source
software. It was launched by the European Commission at the instigation
of the European Parliament after the discovery of the Heartbleed bug in
2014.

Following the success of an initial pilot, the project was renewed for
another three years. This builds on the pilot project by extending the
auditing of free and open source software through:

- setting up bug bounty programmes
- organising hackathons and conferences
- engaging with developer communities