awesome : ajout Open Source Software Capital (OSSC)
"OSS Capital is the first and only seed-stage focused VC exclusively
dedicated to backing leading COSS (commercial open source) startup
founders on a global basis.

Through evangelizing the benefits of COSS in every business category, we
aim to accelerate global knowledge and appreciation of open source
fundamentals at the core of every digital product in existence. If we
are successful, we hope to inspire 1B+ humans to participate in and
create transformative open source technologies by 2030."