awesome : ajout Best Practices for OSS Developers working group
Our objective is to provide open source developers with best practices
recommendations, and with an easy way to learn and apply them.

Unlike other existing best practices list, we want it to be widely
distributed to open source developers and community-sourced. And we want
these practices to stick, thanks to an effective learning platform.