ressources : ajout Open Source Technology Improvement Fund (OSTIF)
The Open Source Technology Improvement Fund is a corporate non-profit
dedicated to securing open source apps that we all depend on. Securing
software isn’t easy, and we know what it takes to succeed. By
facilitating security audits and reviews, OSTIF makes it easy for
projects to significantly improve security.