ressources : open science ajout INOSC
ajout International Network of Open Science & Scholarship Communities

"We believe that the research community itself is best suited to lead
change – those who put OS to practice. That is why local bottom-up Open
Science Communities are crucial in this transition. OSCs engage and
enable academics to make their workflows more open, and allow them to
articulate the opportunities and obstacles they experience, to policy
makers and providers of research infrastructure and services."