ressources : ajout journal Education Ouverte et Libre - Open Education
Education Ouverte et Libre - Open Education is a multilingual scientific
journal that aims to reflect, discuss and think about Open Education
issues at the international level. Following the UNESCO (2021)
recommendations on open science and particularly the "open dialogue with
other knowledge systems", the journal offers a space to reflect on all
aspects of the educational ecosystem in an open and free paradigm.