ressources : Deconstructing Software copyright, 30 years of bad logic
« For over thirty years, software copyright has been a succession of
court cases and law review articles based on bad law, bad logic, bad
mathematics, and/or bad physics (Benson, CONTU, Whelan and Altai being
all of these). I have decided to write a critical review arguing that
software copyright (and dependents like TRIPS, GPL, Bernstein, Junger)
should be abolished in light of 17 USC 102b and its equivalents - for
one reason - it is bad law with no logical basis in the mathematics and
physics of information processing.

What follows is a list of over 90 Acts, decisions and law review
articles I will be critiquing in the review. A small number of cases are
from Asia and Europe, which inherited the problems of the US cases. »