ressources : ajout OERizona Network
The OERizona Network is a professional learning community committed to
providing learners across Arizona and beyond unimpeded and equitable
access to high-quality learning materials through the promotion of open
educational practices. The network seeks to cultivate a vibrant
community of practitioners and advocates in open education by supporting
coordination and collaboration on open educational resource projects and
applications of open pedagogies.