ressources : ajout open education network + open textbook library
Dave Ernst started the Open Textbook Library (OTL) in 2012 at the
University of Minnesota to make finding open textbooks easier for
faculty, and launched an accompanying workshop encouraging them to
consider open textbooks for their courses. His interest and work in
expanding the use of open textbooks was a direct response to escalating
textbook costs in higher education and their negative impact on
students’ academic success, especially on those who are most financially
vulnerable.

In 2014, as more institutions became interested in the OTL and workshop,
Dave announced the creation of the Open Textbook Network to provide a
space for the institutions to talk with each other about open education.
In 2020, the Open Textbook Network became the Open Education Network
(OEN), a new name that better reflects the broad scope of the community.
Today, the OEN is an open education community of both vision and
practice, and the leadership, actions, and results of the community are
transforming higher education.