ressources : ajout OSPO alliance
The OSPO Alliance was launched in June 2021 by European non profit
organisations — OW2, Eclipse Foundation, OpenForum Europe, and
Foundation for Public Code — and concerned individuals to promote an
approach to excellence in open source software management. Together we
created the OSPO Alliance — an open experience-sharing platform to
facilitate discovery of tools and best practices and help define the
state of the art in this domain.