ressources : ajout The Turing Way handbook
« The Turing Way handbook to reproducible, ethical and collaborative data
science.

The Turing Way project is open source, open collaboration, and
community-driven. We involve and support a diverse community of
contributors to make data science accessible, comprehensible and
effective for everyone. Our goal is to provide all the information that
researchers and data scientists in academia, industry and the public
sector need to ensure that the projects they work on are easy to
reproduce and reuse. »