Ressources : ajout International Association for the Study of the Commons
The IASC is the leading professional association dedicated to the
commons. The association, founded in 1989, is devoted to bringing
together multi-disciplinary researchers, practitioners, and policymakers
for the purpose of improving governance and management, advancing
understanding, and creating sustainable solutions for commons,
common-pool resources, or any other form of shared resource.