Project renaming: switch « brick » to « base [of knowledge] » [part 1]
After using brick of knowledge to name this artifact, switch to a base
of knowledge on open models, with the aim to have a more name that
sounds more natural.

It was unclear how to name it at the origin as there is few example from
which getting some inspirations, échec du pavé dans la marre.

This notion of knowledge base may be a rising topics to structure
knowledge, and therefore for openness movements. It will be a way to
promote this kind of artifact, while making it easier to transmit,
knowledge base can be more intuitive to understand and to share.